+++ src/Microsoft.ML.FastTree/Dataset/SegmentIntArray.cs
@@ -565,7 +565,7 @@ namespace Microsoft.ML.Trainers.FastTree
         }
 #endif // USE_FASTTREENATIVE
 
-        internal override void Sumup(SumupInputData input, FeatureHistogram histogram)
+        public override void Sumup(SumupInputData input, FeatureHistogram histogram)
         {
             using (Timer.Time(TimerEvent.SumupSegment))
             {
